// src/lib/service/api-client/model/appResponseData.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * 핀플로우 프로모션: Admin API
 * 핀플로우 프로모션: Admin API 문서
 * OpenAPI spec version: 1.0
 */

/**
 * 앱 정보 Response Data
 */
export interface AppResponseData {
  /** 앱 ID */
  appId?: string;
  /** 앱 이름 */
  appName: string;
  /** 리워드지급을 위한 postbackUrl */
  postbackUrl?: string;
  /** 리워드지급을 위한 hashKey<br/>postbackUrl 호출시 해시키로 해시된 정보를 함께 전송 */
  hashKey?: string;
  /** 설명 */
  description?: string;
}